Attempt Working Example
This is an attempt at using pbcommandR to create a tool that will take
a list of conditions and produce a density plot showing accuracy by
condition type.  I generated it by copying the exampleReseqConditions.R
and modifying the code to do the following things which I considered
examples of needed functionality.

* Load a non-CRAN package
* Parse the conditions contract
* Produce graphical displays.

I have not tested if it works yet, so assume it doesn’t.  Committing
now to make it easier to talk about the code.