Add Simplification Layers for Users
This commit adds wrapper layers which make it easier to write tools
which take a condition table and output a JSON report.  It masks all
the JSON/Tool-Contract/FileType issues from the user, and implements
some “Sugar” functions to output tables and plots.

This is an experimental first step that I will use, if I find I like
the design I intend to build a vignette around it.